SRSampling: take one or multiple simple random samples

SRSampling: take one or multiple simple random samples

SRSampling(
  x,
  size,
  replace = FALSE,
  rep = 1,
  show.SRS = FALSE,
  verbose = TRUE,
  na.rm = FALSE,
  trim = 0,
  ...
)

Arguments

x
the assumed population from which the SRS are chosen. It must be a vector with length > 1. NA in x can also be chosen.

size
the sample size. the number of items to choose

replace
whether sampling should be with replacement

rep
the number of simulation runs. Default rep=1.

show.SRS
whether to show SRS for each run of simulations

verbose
whether to return verbose message

na.rm
whether to remove NA when calculate summary statistics for SRS.

trim
the fraction (0 to 0.5) of observations to be trimmed from each end of x before the mean is computed. Values of trim outside that range are taken as the nearest endpoint

...
additional arguments to be passed to or from methods.

Value

a list a list containing the following components, conditional on the argument rep: When rep=1 (the default):

sample.mean: a numeric number giving the mean of the SRS

sample.se: a numeric number giving the SD of the SRS

SRS: a matrix containing the values in the SRS with nrow=1 and ncol=size. Available only if show.SRS=TRUE

When rep>1:

sample.mean: a list containing the following components about sample means from multiple SRS:

values: a vector with a length of rep giving sample mean for each of the multiple SRS

mean: a numeric number giving the mean of values

se: a numeric number giving the mean of values

sample.se: a list containing the following components about sample sds from multiple SRS:

values: a vector with a length of rep giving sample sd for each of the multiple SRS

mean: a numeric number giving the mean of values

se: a numeric number giving the mean of values

SRS: a matrix containing the values in each of the multiple SRS with nrow=rep and ncol=size. Available only if show.SRS=TRUE.

Examples

pop.1 <- rnorm(1000,0,10)
# take a SRS from x of size=10, show.SRS=TRUE
SRSampling(x=pop.1, size=10, show.SRS=TRUE)

#> A simple random sample of size 10 is drawn from x
#> 
#> The SRS for each run of samplings are shown in the result
#> 

#> $sample.mean
#> [1] 4.703076
#> 
#> $sample.se
#> [1] 11.22522
#> 
#> $SRS
#>          [,1]      [,2]      [,3]     [,4]      [,5]      [,6]      [,7]
#> [1,] 18.07087 0.3849955 -2.166291 4.768438 -6.570574 -9.974431 -3.470885
#>          [,8]     [,9]    [,10]
#> [1,] 8.767773 12.41263 24.80823
#> 

# take 20 SRS from x of size=10, show.SRS=FALSE
SRSampling(x=pop.1, size=10, rep=20)

#> 20 runs of simple random sample of size 10 are drawn from x
#> 
#> The SRS for each run of samplings are NOT included in the result
#> 

#> $sample.mean
#> $sample.mean$values
#>  [1] -10.5992995  -2.4645059   3.6621685   3.8722055   0.2039372  -0.3420757
#>  [7]   0.7356245   4.0024178  -3.4364846  -0.1411582   1.1365655  -2.8541025
#> [13]   4.2445203   3.4185136   2.4451474   0.5074188  -2.6448420  -1.2266083
#> [19]   5.8116487   0.8172250
#> 
#> $sample.mean$mean
#> [1] 0.3574158
#> 
#> $sample.mean$se
#> [1] 3.707296
#> 
#> 
#> $sample.se
#> $sample.se$values
#>  [1] 11.939111 12.067874 12.455150 12.793271 10.214753  8.293052  9.524084
#>  [8] 13.522999  8.850345  7.068467 10.287013  6.353818  7.823334 12.781816
#> [15]  7.545748 10.582422 10.265528  9.853843 11.922616  6.719497
#> 
#> $sample.se$mean
#> [1] 10.04324
#> 
#> $sample.se$se
#> [1] 2.22555
#> 
#> 

# size>length(pop.2)
pop.2 <- rnorm(10,0,10)
# SRSampling(x=pop.2, size=15, rep=20) # error

# non-integer size
SRSampling(x=pop.2, size=5.8) # warning

#> Warning: size=5.8 is not an integer, and ceiling() function was applied.
#>  size=6 was used

#> A simple random sample of size 6 is drawn from x
#> 
#> The SRS for each run of samplings are NOT included in the result
#> 

#> $sample.mean
#> [1] -0.8715857
#> 
#> $sample.se
#> [1] 8.114037
#> 

# x=numeric(0)
# SRSampling(x=numeric(0), size=1) # error

# x with NA
pop.3 <- c(1,2,3,4,NA)
SRSampling(x=pop.3, size=5, na.rm=TRUE)

#> Warning: 
#> size is equal to the length of x

#> A simple random sample of size 5 is drawn from x
#> 
#> The SRS for each run of samplings are NOT included in the result
#> 

#> $sample.mean
#> [1] 2.5
#> 
#> $sample.se
#> [1] 1.290994
#> 

SRSampling(x=pop.3, size=5)

#> Warning: 
#> size is equal to the length of x

#> A simple random sample of size 5 is drawn from x
#> 
#> The SRS for each run of samplings are NOT included in the result
#> 

#> $sample.mean
#> [1] NA
#> 
#> $sample.se
#> [1] NA
#>